Make minor formatting changes.
I've made a couple of minor fomatting changes to white space, header levels, and example texts.
I've changed the title header to not be the tools slugified name. This will make it more readable. It now referes to the package as well. For this commit it refers to the CLI. The anchor link to this header has also been updated to reflect the title change.
Language / highlight identifiers have also been added to code blocks.
URLs to internal code files have also been updated to use relative url instead of hard coded version URLs.