Complete refactor
The README wasn't particularly helpful as it was. I removed the TODOs and added them as GitHub issues.
I added a simple one line description of this package as well as changing the title to match the format that was used in the CLI package README.
This commit is meant to be an initial stating ground for a more comprehensive README.